Refactor structure of tests
