+++ tests/pyside/test_pyside_renderer.py
